+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Attachment, template and HTML behaviour identical to existing email integration
   - Emails using `smtp` support `Importance`,`Priority`,`X-Priority` and `X-MSMail-Priority` headers. Values are mapped automatically from the Supernotify priority.
   - `Message-Id` in email header ties back to notification ID and delivery name
+  - `default_title` option for emails without titles set in notification
 - `delivery_selection` of `fixed` overrides priority and condition filtering at delivery level, if selection is fixed, only `enabled` at transport or delivery level will override your command.
 
 ## 1.16.6